+++ packages/react-doctor/src/cli/colorize-by-score.ts
@@ -1,5 +1,5 @@
 import { SCORE_GOOD_THRESHOLD, SCORE_OK_THRESHOLD } from "../constants.js";
-import { highlighter } from "./highlighter.js";
+import { highlighter } from "../core/highlighter.js";
 
 export const colorizeByScore = (text: string, score: number): string => {
   if (score >= SCORE_GOOD_THRESHOLD) return highlighter.success(text);
