+++ packages/react-doctor/src/cli/install-skill.ts
@@ -4,8 +4,8 @@ import { fileURLToPath } from "node:url";
 import { installSkillsFromSource, SKILL_MANIFEST_FILE, type SkillAgentType } from "agent-install";
 import { SKILL_NAME } from "../constants.js";
 import { detectAvailableAgents } from "./detect-agents.js";
-import { highlighter } from "./highlighter.js";
-import { logger } from "./logger.js";
+import { highlighter } from "../core/highlighter.js";
+import { logger } from "../core/logger.js";
 import { prompts } from "./prompts.js";
 import { spinner } from "./spinner.js";
 import { toDisplayName } from "./to-display-name.js";
