+++ packages/react-doctor/src/cli/select-projects.ts
@@ -4,8 +4,8 @@ import {
   discoverReactSubprojects,
   listWorkspacePackages,
 } from "../core/detection/discover-project.js";
-import { highlighter } from "./highlighter.js";
-import { logger } from "./logger.js";
+import { highlighter } from "../core/highlighter.js";
+import { logger } from "../core/logger.js";
 import { prompts } from "./prompts.js";
 
 export const selectProjects = async (
