+++ packages/react-doctor/src/core/format-error-chain.ts
