+++ packages/react-doctor/src/core/highlighter.ts
